app_2/requirements.txt
paho-mqtt
python-dotenv
influxdb-client